requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
bcrypt==4.1.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
fastapi==0.110.0
h11==0.14.0
idna==3.6
ipdb==0.13.13
ipython
jedi==0.19.1
Mako==1.3.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
python-multipart==0.0.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.36.3
traitlets==5.14.2
typing_extensions==4.10.0
uvicorn==0.29.0
wcwidth==0.2.13
